requirements.txt
APScheduler==2.1.2
awsebcli==3.12.1
beautifulsoup4==4.6.0
blessed==1.14.2
boto==2.48.0
botocore==1.8.15
bz2file==0.98
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.12
Flask-KVSession==0.6.2
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
funcsigs==1.0.2
futures==3.2.0
gensim==2.3.0
google-api-python-client==1.5.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.3
MarkupSafe==0.23
mysql-connector==2.1.6
MySQL-python==1.2.5
numpy==1.13.1
oauth2client==2.2.0
oauthlib==2.0.6
pathspec==0.5.0
pyasn1==0.3.5
pyasn1-modules==0.1.4
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
rauth==0.7.3
requests==2.9.1
requests-oauthlib==0.8.0
rsa==3.4.2
run==0.2
scipy==0.19.1
semantic-version==2.5.0
simplejson==3.11.1
simplekv==0.10.0
six==1.11.0
smart-open==1.5.3
SQLAlchemy==1.1.14
tabulate==0.7.5
termcolor==1.1.0
tzlocal==1.5.1
uritemplate==0.6
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.11.15
